fixed the issue
The previous commit introduced unexpected behavior due to the absolute
positioning of the box::after and box::before elements and their top and
left properties. The elements were not centered around the 'box-content'
class, causing the box to overlap with the one behind it and disrupting
the shiny border effect.

To resolve this issue, I have adjusted the positioning by adding the
following formulas as comments:

padding: 1px;

// left = -padding + 1
left: 0px;

// top = -padding
top: -1px;

This commit addresses the issue and ensures proper alignment of the
elements.